requirements.txt
pytorch==1.8.1
transformers==4.1.1
pyTorch_lightning==1.3.0